cleanup: remove tmpnam
This fixes #3318.

tmpnam generates linker warnings.
