+++ google/cloud/storage/CMakeLists.txt
@@ -291,6 +291,8 @@ if (BUILD_TESTING)
         testing/mock_http_request.h
         testing/mock_http_request.cc
         testing/retry_tests.h
+        testing/random.h
+        testing/random.cc
         testing/storage_integration_test.h
         testing/storage_integration_test.cc
         testing/temp_file.h
