+++ google/cloud/storage/benchmarks/BUILD
@@ -25,6 +25,7 @@ cc_library(
     deps = [
         "//google/cloud/storage:nlohmann_json",
         "//google/cloud/storage:storage_client",
+        "//google/cloud/storage:storage_client_testing",
         "@boringssl//:crypto",
         "@boringssl//:ssl",
         "@com_github_curl_curl//:curl",
@@ -45,6 +46,7 @@ load(":storage_benchmark_programs.bzl", "storage_benchmark_programs")
         ":storage_benchmarks",
         "//google/cloud/storage:nlohmann_json",
         "//google/cloud/storage:storage_client",
+        "//google/cloud/storage:storage_client_testing",
         "@boringssl//:crypto",
         "@boringssl//:ssl",
         "@com_github_curl_curl//:curl",
