+++ google/cloud/storage/benchmarks/CMakeLists.txt
@@ -53,6 +53,7 @@ if (BUILD_TESTING)
             ${target}
             PRIVATE storage_benchmarks
                     storage_client
+                    storage_client_testing
                     google_cloud_cpp_common
                     CURL::libcurl
                     Threads::Threads
