+++ google/cloud/storage/storage_client_testing.bzl
@@ -22,12 +22,14 @@ storage_client_testing_hdrs = [
     "testing/mock_fake_clock.h",
     "testing/mock_http_request.h",
     "testing/retry_tests.h",
+    "testing/random.h",
     "testing/storage_integration_test.h",
     "testing/temp_file.h",
 ]
 
 storage_client_testing_srcs = [
     "testing/mock_http_request.cc",
+    "testing/random.cc",
     "testing/storage_integration_test.cc",
     "testing/temp_file.cc",
 ]
